User&purchases

+++ database/migrations/2020_08_27_060023_create_products_table.php
@@ -18,8 +18,8 @@ class CreateProductsTable extends Migration
             $table->string('name');
             $table->string('description');
             $table->string('ref_img');
-            $table->float('number_of_pieces');
-            $table->integer('price');
+            $table->integer('number_of_pieces');
+            $table->float('price');
             $table->timestamps();
         });
     }
